refactor: 이름 중복 검증 책임을 유틸 클래스로 이동
